Update PrevisionController to pull real amounts for année and adjust totals; align frontend to show dynamic realisations and remove zero-only alerts. Also minor frontend tweak to remove empty alert block for no previsions.

X-Lovable-Edit-ID: edt-23624b18-4442-4ded-8f7d-0a959c83f52b

+++ backend/app/Http/Controllers/Api/PrevisionController.php
@@ -141,19 +141,19 @@ class PrevisionController extends Controller
         $reelBanque = $this->getReelsBanqueAnnee($annee);
 
         // Prévisions enregistrées
-        $prevuRecettesCaisse = Prevision::where('type', 'recette')
+        $prevuRecettesCaisse = (float) Prevision::where('type', 'recette')
             ->where('source', 'caisse')
             ->where('annee', $annee)
             ->sum('montant_prevu');
-        $prevuRecettesBanque = Prevision::where('type', 'recette')
+        $prevuRecettesBanque = (float) Prevision::where('type', 'recette')
             ->where('source', 'banque')
             ->where('annee', $annee)
             ->sum('montant_prevu');
-        $prevuDepensesCaisse = Prevision::where('type', 'depense')
+        $prevuDepensesCaisse = (float) Prevision::where('type', 'depense')
             ->where('source', 'caisse')
             ->where('annee', $annee)
             ->sum('montant_prevu');
-        $prevuDepensesBanque = Prevision::where('type', 'depense')
+        $prevuDepensesBanque = (float) Prevision::where('type', 'depense')
             ->where('source', 'banque')
             ->where('annee', $annee)
             ->sum('montant_prevu');
